Refactored like everything. So it is not stateful anymore.

+++ src/main/java/de/majaf/voci/entity/repo/InvitationRepository.java
@@ -3,5 +3,5 @@ package de.majaf.voci.entity.repo;
 import de.majaf.voci.entity.Invitation;
 import org.springframework.data.repository.CrudRepository;
 
-public interface InvitationRepository extends CrudRepository<Invitation, Integer> {
+public interface InvitationRepository extends CrudRepository<Invitation, Long> {
 }
