+++ src/main/java/de/majaf/voci/entity/repo/RoomRepository.java
@@ -3,5 +3,5 @@ package de.majaf.voci.entity.repo;
 import de.majaf.voci.entity.Room;
 import org.springframework.data.repository.CrudRepository;
 
-public interface RoomRepository extends CrudRepository<Room, Integer> {
+public interface RoomRepository extends CrudRepository<Room, Long> {
 }
