+++ src/main/java/de/majaf/voci/entity/repo/TextChannelRepository.java
@@ -3,5 +3,5 @@ package de.majaf.voci.entity.repo;
 import de.majaf.voci.entity.TextChannel;
 import org.springframework.data.repository.CrudRepository;
 
-public interface TextChannelRepository extends CrudRepository<TextChannel, Integer> {
+public interface TextChannelRepository extends CrudRepository<TextChannel, Long> {
 }
