+++ src/main/java/de/majaf/voci/entity/repo/VoiceChannelRepository.java
@@ -3,5 +3,5 @@ package de.majaf.voci.entity.repo;
 import de.majaf.voci.entity.VoiceChannel;
 import org.springframework.data.repository.CrudRepository;
 
-public interface VoiceChannelRepository extends CrudRepository<VoiceChannel, Integer> {
+public interface VoiceChannelRepository extends CrudRepository<VoiceChannel, Long> {
 }
